pyproject.toml
[tool.poetry]
name = "insighter"
version = "0.1.0"
description = ""
authors = ["artemprod <artemprod1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
docker = "^7.0.0"
environs = "^11.0.0"
aio-pika = "^9.4.1"
uvicorn = "^0.29.0"
fastapi = "^0.110.2"
pydantic = "^2.7.1"
pika = "^1.3.2"
nats-py = "^2.7.2"
pydub = "^0.25.1"
pytube = "^15.0.0"
openai = "^1.25.0"
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

